requirements.txt
numpy
Pillow
scikit-image
scipy
beautifulsoup4
torch
torchvision